Additions/x11/vboxvideo: prevent a potential integer overflow.
svn:sync-xref-src-repo-rev: r99796

+++ src/VBox/Additions/x11/vboxvideo/vboxvideo.c
@@ -1,4 +1,4 @@
-/* $Id: vboxvideo.c 55448 2015-04-27 13:10:59Z noreply@oracle.com $ */
+/* $Id: vboxvideo.c 55450 2015-04-27 13:26:10Z noreply@oracle.com $ */
 /** @file
  *
  * Linux Additions X11 graphics driver
@@ -282,8 +282,8 @@ static Bool adjustScreenPixmap(ScrnInfoPtr pScrn, int width, int height)
                        adjustedWidth, height, (unsigned) pVBox->cbFBMax / 1024);
             return FALSE;
         }
-        vbvxClearVRAM(pScrn, ((size_t)pScrn->virtualX) * pScrn->virtualY * pScrn->bitsPerPixel / 8,
-                      adjustedWidth * height * pScrn->bitsPerPixel / 8);
+        vbvxClearVRAM(pScrn, ((size_t)pScrn->virtualX) * pScrn->virtualY * (pScrn->bitsPerPixel / 8),
+                      ((size_t)adjustedWidth) * height * (pScrn->bitsPerPixel / 8));
         pScreen->ModifyPixmapHeader(pPixmap, adjustedWidth, height, pScrn->depth, pScrn->bitsPerPixel, cbLine, pVBox->base);
     }
     pScrn->displayWidth = pScrn->virtualX = adjustedWidth;
@@ -1411,7 +1411,7 @@ static void VBOXLeaveVT(ScrnInfoPtr pScrn)
         vbox_crtc_dpms(pVBox->pScreens[i].paCrtcs, DPMSModeOff);
 #endif
     vboxDisableVbva(pScrn);
-    vbvxClearVRAM(pScrn, ((size_t)pScrn->virtualX) * pScrn->virtualY * pScrn->bitsPerPixel / 8, 0);
+    vbvxClearVRAM(pScrn, ((size_t)pScrn->virtualX) * pScrn->virtualY * (pScrn->bitsPerPixel / 8), 0);
 #ifdef VBOX_DRI_OLD
     if (pVBox->useDRI)
         DRILock(xf86ScrnToScreen(pScrn), 0);
@@ -1442,7 +1442,7 @@ static Bool VBOXCloseScreen(ScreenPtr pScreen)
             vbox_crtc_dpms(pVBox->pScreens[i].paCrtcs, DPMSModeOff);
 #endif
         vboxDisableVbva(pScrn);
-        vbvxClearVRAM(pScrn, ((size_t)pScrn->virtualX) * pScrn->virtualY * pScrn->bitsPerPixel / 8, 0);
+        vbvxClearVRAM(pScrn, ((size_t)pScrn->virtualX) * pScrn->virtualY * (pScrn->bitsPerPixel / 8), 0);
     }
 #ifdef VBOX_DRI
 # ifndef VBOX_DRI_OLD  /* DRI2 */
